requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.3
pydantic==1.10.2
python-dotenv==1.0.0
typing_extensions==4.4.0
yarl==1.8.2
